Refactor: Move modules to `ui` and `services`
